like service rework

+++ Smakoowa-Api/Smakoowa-Api/Services/LikeService.cs
@@ -1,36 +1,31 @@
 ﻿namespace Smakoowa_Api.Services
 {
-    public abstract class LikeService
+    public abstract class LikeService<T> where T : Like
     {
-        protected readonly IHelperService<LikeService> _helperService;
-        protected readonly ILikeRepository<Like> _likeRepository;
+        protected readonly IHelperService<LikeService<T>> _helperService;
+        protected readonly IBaseRepository<T> _likeRepository;
         protected readonly IApiUserService _apiUserService;
         protected readonly ILikeValidatorService _likeValidatorService;
+        protected readonly LikeableType _likeableType;
 
-        public LikeService(ILikeRepository<Like> likeRepository, IHelperService<LikeService> helperService, IApiUserService apiUserService,
-            ILikeValidatorService likeValidatorService)
+        public LikeService(IBaseRepository<T> likeRepository, IHelperService<LikeService<T>> helperService, IApiUserService apiUserService,
+            ILikeValidatorService likeValidatorService, LikeableType likeableType)
         {
             _likeRepository = likeRepository;
             _helperService = helperService;
             _apiUserService = apiUserService;
             _likeValidatorService = likeValidatorService;
+            _likeableType = likeableType;
         }
 
-        protected async Task<ServiceResponse> AddLike(Like like)
+        public async Task<ServiceResponse> RemoveLike(int likedId)
         {
-            try
-            {
-                await _likeRepository.Create(like);
-                return ServiceResponse.Success("Like added.");
-            }
-            catch (Exception ex)
-            {
-                return _helperService.HandleException(ex, "Something went wrong while adding a like.");
-            }
-        }
+            var likeToRemove = await _likeRepository.FindByConditionsFirstOrDefault(
+                c => c.LikedId == likedId
+                && c.CreatorId == _apiUserService.GetCurrentUserId());
 
-        protected async Task<ServiceResponse> RemoveLike(Like likeToRemove)
-        {
+            if (likeToRemove == null) return ServiceResponse.Error($"Like of item with id: {likedId} not found.");
+
             try
             {
                 await _likeRepository.Delete(likeToRemove);
@@ -39,6 +34,47 @@
             catch (Exception ex)
             {
                 return _helperService.HandleException(ex, "Something went wrong while removing a like.");
+            }
+        }
+
+
+        public async Task<int> GetLikeCount(int likedId)
+        {
+            return (await _likeRepository.FindByConditions(rl => rl.LikedId == likedId)).Count();
+        }
+
+        public async Task<ServiceResponse> AddLike(int likedId)
+        {
+            var validationResult = await _likeValidatorService.ValidateAddLike(likedId);
+            if (!validationResult.SuccessStatus) return validationResult;
+
+            var newLike = new Like { LikedId = likedId, LikeableType = _likeableType };
+
+            try
+            {
+                await _likeRepository.Create((T)CreateNewLike(likedId, _likeableType));
+                return ServiceResponse.Success("Like added.");
+            }
+            catch (Exception ex)
+            {
+                return _helperService.HandleException(ex, "Something went wrong while adding a like.");
+            }
+        }
+
+        private Like CreateNewLike(int likedId, LikeableType likeableType)
+        {
+            switch (likeableType)
+            {
+                case LikeableType.Recipe:
+                    return new RecipeLike { LikeableType = likeableType, LikedId = likedId };
+                case LikeableType.Tag:
+                    return new TagLike { LikeableType = likeableType, LikedId = likedId };
+                case LikeableType.RecipeComment:
+                    return new RecipeCommentLike { LikeableType = likeableType, LikedId = likedId };
+                case LikeableType.CommentReply:
+                    return new CommentReplyLike { LikeableType = likeableType, LikedId = likedId };
+                default:
+                    return null;
             }
         }
     }
